♻ 코드 수정: RecommendedWorkoutHistory VO,DTO routineRatings 타입 변경

+++ healthtart/src/main/java/com/dev5ops/healthtart/recommended_workout_history/domain/dto/RecommendedWorkoutHistoryDTO.java
@@ -16,7 +16,7 @@ public class RecommendedWorkoutHistoryDTO {
     private Long historyCode;
 
     @JsonProperty("routine_ratings")
-    private int routineRatings;
+    private Double routineRatings;
 
     @JsonProperty("recommended_time")
     private int recommendedTime;
